Introduce a slight delay in changing the min height of an output area that is cleared to prevent jitter.
When an output area is cleared and then quickly replaced with other output, the quickly changing height causes the page to jitter. This happens a lot with ipywidget @interact, for example. This 50ms delay seems small enough to take care of the problem, but could be tweaked.

An example that shows this jitter is this:

```python
from ipywidgets import interact, Output, HBox
left = Output()
right = Output()
with left:
    @interact(x=(0,10))
    def f(x):
        @interact(y=(0,x))
        def g(y):
            with right:
                print('Product',x*y)

print("Product calculator")
HBox([left, right])
```