Improve scroll behavior for active cells.
In this commit, we decouple setting a cell as active and scrolling to show it. This is useful when, for example, you are making a cell active by clicking on it. Clicking indicates a point of interest in the cell (like a position in the input editor, or a particular output slider you want to interact with), and scrolling when these click events make a cell active disrupts the user’s focus on the part they want to interact with. On the other hand, programmatically setting the active cell often does involve scrolling, so we added a method to explicitly scroll to the active cell.

Fixes #478. Also fixes the same long-standing problem with the current notebook.