CSS workaround for sidebars on Safari
Safari's flexbox bug appears to miscalculate
the stretch window for the side bars,
causing the contents of the vertical bars to be shrunk to the point of being invisible.

Setting a minimum width (height) works around this miscalculation,
but is probably not the right thing to do in the long run.