renderers: No margin for second level lists
There was a margin put in lists to separate it from paragraphs,
header, and other tags in general. But this margin is also causing a
space in the second level list if the first level list item has a html
tag. Hence, we add a new CSS selector which sets the margin in this case
back to 0.