Add Typescript Sourcemaps
Create sourcemaps when compiling TS files so that you can see
the original typescript source when you debug the app. This requires
a more resource intensive sourcemap creation in webpack, so we don't
enable it by default. On my laptop, enabling it increased build times
from ~9s to ~13s..

Fixes #506