Focus the text editor intelligently.
We focus the actual text editor when the tab is clicked, or when there is a mouse click anywhere in the entire widget when the text editor isn’t focused already. This lets us do things like switch tabs back and forth with keyboard shortcuts or clicking on tabs to pick up the cursor where we left off, or click on the line number portion of the editor to get the cursor focused in the chosen editor.

This extends the work done in #3278.