Handle theme loading errors (#3298)
* Add error handling to the theme manager load css function.

Fixes #3295

Previously, if loading a css file had an error, the theme manager would never be ready. Now the theme loader recognizes there is an error, logs the error, and continues.

* Load the optional roboto fonts outside of the css.

This way an error in loading the optional font (for example, on a computer note connected to the network) doesn't log an error.