Correctly match an output area model index to the layout children index.
In particular, we must skip inputs, which are not represented in the model.

In this implementation, we essentially count up the children widgets that we see that are not inputs until we get to the correct index, or we get to the length of the widgets array.

Fixes the issue pointed out in https://github.com/jupyterlab/jupyterlab/pull/3514#issuecomment-354885116