Make notebook some context menus global (#3765) (#3909)
Currently, "Clear All Outputs", "Enable Scrolling for Outputs", and
"Disable Scrolling for Outputs" context menus only show up
when you right click on a code cell. This changes them to show up when
you right click anywhere in the notebook.

It does this by making them global and re-ordering those options
so that they show up between the existing global menu items and the
current cell menu items. This preserves the grouping of "Clear Outputs"
and "Clear All Outputs" when a code cell is selected.