Make the output clone a simplified output widget without prompts. (#3730)
When cloning an output area, we don’t want to handle input events because we want only one source for the input.

Also, we wanted to simplify the output area DOM structure so that individual rendered outputs could set themselves to grow (using flexbox), measure themselves after growing, so they could resize themselves to possibly take up the entire area. For example, we might have an ipywidget that wants to grow when in a separate tab, so it can listen to a resize event, set a flex-grow attribute, and measure the size to relayout.