Add parsing options to return row offsets, or column offsets, as well as a max number of rows or set number of columns.
This sets the stage for incrementally parsing the file. Current idea is to initially just parse for row offsets. That also gives us the number of rows and columns. We decide based on the size of rows*cols whether we want to create an offset array that is rows*cols, or just rows. If we decide we have too many fields for a full rows*cols array, we just store rows and parse as needed. If we have a rows*cols array, we parse rows as we are asked for the data in the rows, and incrementally fill in the offset array (i.e., treat the offset array as a cache for all field offsets).

Even in the case of just storing row offsets, we should maintain a cache of column offsets for, say, 500-1000 rows, invalidated when we ask for any data item outside the cache. That should allow for full-screen paints and sideways scrolling without reparsing the data.