Make extension install command find compatible version (#3608)
* Make extension install compatible version

* Update semver file

* Add test + fix issues for compatible extension install

* Cleanup debugging

* Refactor + fix fetching of package metadata

- Move function out of class
- Ensure scoped packages work
- Move exception handling up, handle wider exception type

* Extract parts of _test_overlap for comparison

_compare_ranges gives -1, 0, 1, specifying which direction the ranges
are compared to each other if the do not overlap.

* Don't pass tempdir to _install_extension

Where _install_extension unpacks seems as in internal detail. If
reverting, ensure a new tempdir is passed in recursive call.

* Remove debug print

* Move warning to avoid double-printing

* More helpful error for incompatible versions

* Check whether older version is valid

When searching for the latest compatible package, ensure a version that
satisfies the dependencies are also valid. If not we can end up with
really old package versions that were simply not declared correctly (or
following older guidlines).

* Do pass tempdir to _install_extension

But make sure that recursive call gets a clean dir.

* Fix test patch