Delete the regex processing option for searching for the end of an unquoted field.
In some cases and different browsers, the regex approach below is faster by
quite a bit, but in others, the loop is faster. More testing is needed to see
which approach we want to keep. In my microbenchmarks, it seemed that Firefox
was faster with the plain loop we are keeping in this commit, and Chrome seemed
faster with the regex.

const endfield = new RegExp(`[${delimiter}${rowDelimiter}]`, 'g');

endfield.lastIndex = i;
let match = endfield.exec(data);
if (match) {
  i = match.index;
  char = data.charCodeAt(i);
}