Correct keyboard shortcuts commands for headings
The proposed changes were based on @jasongrout's suggestions in issue #4428.

In `jupyterlab/packages/shortcuts-extension/schema/plugin.json`, keyboard shortcut `1-6` are registered with commands `change-`**`to-cell`**`-heading-X`, whereas the commands in `jupyterlab/packages/notebook-extension/src/index.ts` are named `change-`**`cell-to`**`-heading-X`. As a result, the keyboard shortcuts are not working. 

This PR attempts to fix this problem by changing the command names in `jupyterlab/packages/shortcuts-extension/schema/plugin.json` to the ones in `jupyterlab/packages/notebook-extension/src/index.ts`.