Change running widget to use react
Changes the RunningSessions widget to use react instead of phosphor. The
code is more readable for me after the refactor and more concise.

It also removes the ability to customize the renderer of the running
widget, because I am not sure exactly how this should translate to the
react world and don't see any examples using it in our codebase
(or on github).

I made a common component for the running terminals and sessions, because
they shared a lot of the same UI and code.

I tried to follow the recommendations in the React docs to hold the state
as close to where it is being used. So I stored the models on separate
list components, not all in one big one. I also tried to be cognizant
of render times. I opted not to mess with the `shouldComponentUpdate`
to optimize re-rendering, but since the kernels and the terminals
state is stored in different components, when one is updated the other
is not.