Tighten trust/safety for renderers
A lot of renderers rely on not being asked to render untrsuted models if
they mark themselves as unsafe. This is not currently the case, but it
likely should be. While I cannot find any examples in the wild who
benefit from trying to render an untrusted model with an unsafe
renderer, I've left the door open for this pattern by allowing 'prefer'
to be passed as a value for the preferredMimeType call.