Improve labext link usablility
yarn deduplication sometimes depend on the order that dependencies are
specified in. For this reason, if a library package is linked, and an
extension that depends on it is installed with a local install, the
linked package needs to be specified first in order for correct
deduplication to work. While this should ideally be done intelligently
by dependency resolution, simply putting the linked packages before the
local ones is a simple fix for most cases.