Fix toolbar collapse status based on explicit hiding, rather than cumulative visibility.
We use .isHidden rather than .isVisible since .isVisible is affected by parent visibility. We just want to test if the direct children are explicitly hidden to decide whether to expand or collapse the toolbar. This matters, for example, if trying to fit a widget where some parent isn’t attached or visible yet (for example, a tab that is not shown).