Just use the BoxLayout instead of trying SingletonLayout.
The BoxLayout positions its children, so size information propagates up and down. See https://github.com/phosphorjs/phosphor/issues/340 for more details.

In particular, I was seeing issues with using SingletonLayout in the editor, where the codemirror editor was not understanding it was inside a small window and needed to have scrollbars.