Update the current kernel session id on *every* message.
The kernel session id is stored to make it easy to see if a previously scheduled message should still be processed, so it should be updated every time we get anything from the kernel.

This fixes problems we were seeing where the messages updating the kernel session id were being ignored because they were being canceled.