Don’t preemptively cancel messages from a different kernel session.
This is so that code waiting on, for example, a status message will still resolve, especially in the case that the messages were buffered into another session.

Any message handling that may require the current session (such as creating comms) should check to see if the message being handled is current.