Include chrome-test.js in package
Hopefully fixes https://github.com/jupyterlab/jupyterlab/issues/4873.

I see that `node-version-check.js` *is* included in the package, but `chrome-test.js` is not and they are in the same folder. Only difference is that `node-version-check.js` is part of this `package_data_spec` list but `chrome-test.js` is not. So hopefully adding it there will add it to the package.