Mismatch in apputils and application version
Following the tutorial I ran into a `No provider for: @jupyterlab/apputils..` error. I assume the issue was that the command above installed version `0.17.0-1` as changing to 0.16.0 (aligning with the `"@jupyterlab/application": "^0.16.0",` from the package.json in boilerplate) fixed the issue.

Alternatively the version of `@jupyterlab/application` in the generate boilerplate should match the latest.