Improve installation instructions, fixes #4940 (#4948)
* Improve installation instructions, fixes #4940

The installation instructions assumed that users already
knew exactly what conda, pip, and pipenv were, and that
they already had one of them.  For many potential users
that simply isn't true (especially if they're using
Jupyter for something other than Python).  In addition,
on many systems `pip` means "Python2"; provide guidance
to encourage people to install JupyterLab to run on Python3
(unless they have specific reasons to not do that).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Remove extraneous text from README

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak install instructions further per feedback

This modifies the installation instructions further.
It turns out pip is a little complicated, because we want to
prefer installing on Python version 3 and there are different
ways to request that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix README.md install directions using jlpm run prettier

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Resolve a few issues that were identified by jasongrout

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>