evaluate return code of node-version-check.js to avoid misclassification (#5083)
The process will exit with return_code == 0 if a recent version of node is installed or with return_code == 1 if it is not installed or below version 5.

The previous try/excpet implementation raised the ValueError if there was any problem (including those unrelated to nodejs) which complicates troubleshooting. In my particular case, there was an error caused by Process() due to a deprecated version of subprocess32.