Convert code editor jsoneditor module to not use React
We weren’t really using react philosophy, just using it to render the initial DOM, and this meant we had to do a lot of workarounds to retrieve DOM elements. I think it’s simpler to just convert this back to traditional DOM manipulation unless/until we want to move everything (events, etc.) over to React.