Fixes #5740 by re-rendering button on many-changed
This also uncomments the toolbar command button tests. There was some
mismatch between the testing behavior and the implemented behavior. I
opted to maintain the implemented behavior as much as possible and
change the tests. The only exception was passing the command `label`
in as the `label` prop in `propsFromCommand`. Otherwise, it is impossible
to render a label only command.