Revamp cell tools, slim down the notebook instance tracker.
This adds the basic capability for notebook-wide metadata in the cell tools (now called notebook tools). We also fixed a few bugs dealing with updating the metadata based on the metadata widgets, and made it possible to delete metadata from the select controls as well.

This deletes most of the “save with options” functionality, though we’ll need to still delete more to clean it all up.