Refresh the editor when switching a Markdown cell to edit mode.
In ad2d92278c82fca57eddc32628add597269575eb (#5700), editor refreshing is eliminated in AfterShow messages for cells. This means when opening a notebook with a rendered Markdown, double-clicking to edit the Markdown cell now no longer refreshes the editor, which means the editor initially is the wrong height (often zero height). This changes forces an editor refresh explicitly in this case.