Delete configuration for cell collapse, scroll, and editable metadata syncing.
This forces all cells to synchronize the view and model metadata always for collapse state, output scrolled state, and editable (readonly) state. It removes the notebook configuration options for turning off this syncing, and removes the notebook tools widgets for controlling the syncing or changing the values.

We might introduce these later, but for now we are simplifying to the basic model that a cell’s visual state should directly represent this metadata at all times, which means that this visual state is saved automatically.

I think removing the notebook tools controls also means that we don’t have a good way to set a cell as readonly.