Update communication channels in readme
It was discussed in the Jupyter team meeting in March 2019 about
different communication channels in use. It seemed like there were some
agreement that it can be time inefficient to use Gitter to help users as
the answers would probably not help others.

In this commit I aligned the recommendations on the JupyterLab repo with
JupyterHub/BinderHub/zero-to-jupyterhub-k8s. It involved removing the
link to the Jupyter google group, and adding a badge to the discourse
forum.