Let an output overflow with scrollbars in both directions.
In linked views, CSS at the end of this file was setting the height to 100% if the output was the only child. Without an overflow-y setting, this was hiding content without a scrollbar. You can see it in the following example:

for i in range(1000):
    print(i)

and “Create new view for output” (a linked output cell). With a single output, the vertical scrollbar does not appear in this new view, and we cannot see the bottom of the output.


Rather than just fixing that one case of an only child in a linked output view, we make this change, following the principle that output contents should always be accessible and not hidden.