Remove support for jupyter.outputs_hidden metadata
Do not set jupyter.outputs_hidden until we can fully support it both in reading and writing. See https://github.com/jupyterlab/jupyterlab/pull/5968#issuecomment-477170937 and https://github.com/jupyter/nbformat/issues/137. Basically, it is unclear what to do in case of a conflict between the `collapsed` and `jupyter.outputs_hidden` metadata fields. For backwards compatibility, we clearly need to respect and write to `collapsed`. Since `jupyter.outputs_hidden` is relatively new and unimplemented, let’s wait for clarification on it before supporting it, so we don’t have to change how we support it in the future.