Remove insistence on testing for notebook 4.3.1
IIRC, notebook 4.3.1 is not compatible with tornado 6, so either we need to also pin tornado in our tests, or just let notebook be the most recent version. It’s more relevant to test the most recent version of the notebook and tornado, so let’s remove the testing pin for notebook 4.3.1.