Note that server directories are for information only.
These variables often contain information about how the server is set up, and the server may not expose this information for security or because it does not apply (for example, config may come from a database, not a filesystem path). We need to be very clear to people using these paths that they are not to be relied on in code.

See #5313 and #5402 for more context.

Fixes #5313