Change breadcrumbs to use a folder icon for root and slashes for separators
This implements @tgeorgux’s design from https://github.com/jupyterlab/jupyterlab/issues/6207

The slight difference in this commit is that it also includes a final slash. This adds a bit to the visual complexity, but makes it much more clear that the final thing is a directory, and you should think about the full path as that directory, slash, something in the filebrowser list.