Default to production build
Now that the production build is fixed, I think it's better to default to it. Indeed, we want users installing JupyterLab to have the best experience using the interface and it means running the optimized production build.

This is especially true for extensions built with React because in dev mode, they will use the React's development bundle which is known to be much slower than the production one.

See also: https://github.com/jupyterlab/jupyterlab/issues/6804#issuecomment-510721047