DnD attachments via rich contents mime
Adds a rich contents mime type to allow extensions to more directly use the already fetched content models of the filebrowser. This is then used in the attachment cell widgets to synchronously insert the name (which will initially resolve to a broken image), and then start the async fetch.

Also reefactors `updateCellSourceWithAttachment`.