Client session should dispose the session it owns when it is disposed.
Fixes #7144

This causes an issue with the session manager list of sessions, which should reflect which sessions are active on the server. When a session is disposed, it emits a terminated signal, which in turn causes the session model to be deleted from the session manager’s list, indicating that the session does not exist on the server when it in fact does. We have conflated deleting the js object with the server-side session terminating. The net result of this is that the session in the list of running sessions and the green session indicator on a filename both disappear until the next server session poll, at which point they are restored to accurately reflect what sessions are active on the server.

To solve this issue, we do not automatically delete the session model, but instead refresh our session list from the server, to make sure our session model list accurately reflects what is on the server.