Support for output search
This is the first pass at search which can search output in a generic
way. This commit does this on a best effort through walking and
manipulating the DOM. There are limitations, including the inability to
handle that to a user <span><b>foo</b>bar</span> should be searched as
foobar. This also will not handle SVGs and does not expose a way to
delegate search to output presenters. Still, this works very well on
stdout/stderr type messages, tables, etc.

Fixes: #6768