Change kernel.connectComm to kernel.createComm and add a new hasComm method.
Previously, if you tried to create a comm that already existed, you could get the comm someone else had created, which means you would overwrite their message handlers, etc. This was a bad design decision.

This changes the API so that you can only create comms with ids that do not exist. You can also query to see if a given id is used.