Move the scrolling widget back to the LogConsolePanel.
We have two basic approaches - we can have the scrolling behavior at the MainAreaWidget level (one intersection observer for the entire log console, but harder api since there is another wrapper around the content), or the scrolling behavior at the output area level (inside log console, which means an intersection observer for each output area, but simpler top-level api). I’m opting for the latter, and a bonus is that perhaps the scroll position can be preserved across different logs.

We could also make a new main area widget with the scrolling behavior, but that seems a bit too specialized.