Do the browser default behavior when clicking on a label in a toolbar.
We were too aggressive about activating the parent widget when getting a click in the toolbar. If we have a label for a control, clicking on the label by default focuses the control, which is what we want, but we were often wresting focus away from the control when we called the activate function.