Disable type checking vega-lite libraries.
Typscript 3.7 throws some errors when compiling Vega-lite 2 and 3, so we turn off lib checks.

This was investigated much more deeply in https://github.com/jupyterlab/jupyterlab/pull/7428, where we deleted the vega 4 extension and upgraded the vega 5 extension to vega-lite 4.0b11. The vega-lite upgrade was not going smoothly, so we moved it to a separate issue and redid the Typescript upgrade by just using the skipLibCheck flag.