Adding "Tags" heading to Pull Request Template
I've been working on the Design System with @isabela-pf to complete it and @Zsailer to help integrate it with Developer Workflow. I'm proposing we add to the PR template for people to tag if they've made changes to any .css files so I (or any other Designer) have access to a list of anything that may have updated the design system.

I also thought it would be nice to have the PRs tagged to make it easier for traige.

Let me know what you think.