Teach update-dependency about more range specifiers and make it adopt the current range for any tag.
Currently there is a special case to adopt the existing range specifier if the upgraded range is ‘latest’. This extends this special case for any upgrade to a tag version. This also allows for range specifiers of =, <, >, <=, and >=.

This fixes a bug where updating to ^latest generates a spec of ^^latest in the current version.