Update Periodic Labeler
**Description**
Update periodic labeler so that it labels every PR and only runs every 10 minutes.

**Motiviation**
Previously, the labeler only ran on 30 PRs max, and ran every 10 minutes. The bug was
fixed, so its now likely that it will label every PR. 5 minutes seemed too short of a
period to observe changes to the PR queue, so the period was extended to 10 minutes.