Fixed wrapSelect to support multiple selects
With the new LabIcon updates the wrapSelect changes set the caretdown icon
relative to the module body instead of the select. The inital values made
this work for the one use case in lab, but breaks all uses in extensions.

This pr makes the icon location relative to the select-wrapper div to fix this.