Start adding some docs on workspace format
I am working on using the workspace format to build up dashboards in JupyterLab and struggled a bit to piece together how you go from the JSON specification to opening and tracking windows. So I have put together some initial documentation on it and an example for the layout restorer.

This is my best guess of how things work after looking at the code a bit, but I am not 100%! It could use a review by someone who helped design this system.

Yes, it is under the `urls.rst` doc, sorry, I was lazy and didn't extract out all this to a new file.