Make deduplication commands fail on duplicates
Many places in the code, we assume that the deduplication returns an error code if there were duplicates, but yarn-deduplicate only returns an error code on duplicates if the --fail option was given. Here we give that option everywhere we deduplicate.

This came up when installing a custom widget which supports both JLab 1 and JLab 2, which needs the deduplication strategy to work. An initial installation will not work since yarn was not run after the deduplication. With this change, yarn is run after the deduplication when there were packages deduplicated, and the extension installs correctly.