Make IFuture done typings more correct
The kernel IFuture done promise never returns undefined in practice (since every shell/control message has a reply, which is needed for the done promise to even resolve), so we strengthen the typing in the interface.

This means that things relying on the done resolution do not have to check to see if the value is undefined from a typing perspective.