Don't overwrite existing attachments
Checks if the cell already has an attachment using that name. If yes then start adding numbers such that image.png will become image_1.png etc. The regex is used for this splitting in order to only split on the final . (i.e. allow filenames such as image.image.png) and to include the . for reconstructing the filename.