Set sidebar icon height
Currently icons in the side bar have no height set. This causes the
iconsto use the height in the svg itself. For icons in lab whose
height are hardcoded to 16px this is not an issue, but some extensions
can't edit the height in the svg files and this can cause the icon
height to be very large and cause issues with automated test suites.

This PR sets the `height` of sidebar icons to `auto` which removes
any specific height from the svg.

This PR does not affect any core extensiosn and is solely to better
support extenernal extensions.