Update the singleton packages to include at least every package with a 'tokens.ts' file.
This list is currently manually updated and way out of date. It would
be better to have a way to automatically determine the list of singleton
packages, as this won't remove any packages from the list that may need
to be removed. But this is a step towards at least updating the list.