Lint - updates from the new version of prettier.
I added two prettier config options to minimize the churn - we can remove them when we decide we can deal with the churn they introduce (quite a lot).

The decision to add a space after function names also introduces quite a bit of churn, but there is no option to disable it. See https://github.com/prettier/prettier/pull/3903 (where there was quite a lot of support for adding such an option to minimize needless churn, but was decided against by the prettier team, which was disappointing to me.)