Fix comment explaining the extension entry point.
We used to have an entry point of {}, which meant that no entry point was generated. However, in order to get the public path code to run, we had to add back in an entry point. Webpack may eventually include a module federation option for code to run when the remote entry file is loaded, and then we change this entry point back to {}.

See https://github.com/webpack/webpack/issues/10352#issuecomment-675649389