URLExt.join cant handle colon in relative paths
URLExt.join checks for a colon to seperate protocol prefix from a path.
This happens to work with relative paths because it just sets the prefix
to empty string, but if the relative path includes a colon it thinks
it's a protocol and returns malformed urls.

This fix checks if the given path is relative and sets the prefix to
empty string as it would be if there was no colon.

Fixes #9159