Fix the small offsets in the placement of open menus.
I don’t fully understand everything that is going into the menu position calculation, and in particular our translate to move the menu up so it appears to be continuous with the menu title. There may be a better way to achieve this for both the single-document mode and normal mode. One complication here is that there is an extra border in single-document mode.

One confusing thing to me is the menu panel height is 27px, but the menu items themselves I think are 28px.

Regardless, things look nice now as far as I can tell.