Remove ensure-max-old-space script.
This was a workaround for node 10 that allowed many users to compile JupyterLab without running out of memory (especially if they had lots of extensions installed). It is no longer needed for Node 12 or higher, where memory management is automatic. Node 10 is now in maintenance mode,  (since May 2019) and node 12 is also almost in maintenance mode (starting Nov 2020), so I think we can drop the Node 10 workaround.

Users still on Node 10 will still be able to manually increase their available memory by setting the environment variable NODE_OPTIONS=--max-old-space-size=4096, for example.

This change dropping the workaround for Node 10 was originally done in Dec 2019 in #7677, but then reverted in #8112 (to support node 10 in jlab 2.1).