Set the relative sizes of left/main/right areas in afterAttach.
In #9287, we noticed there seemed to be a race condition when setting the relative sizes on construction, where sometimes those relative sizes were being applied in an update that did not have the proper full width.

Setting the relative sizes in after-attach *may* circumvent this race condition, and make sure that the relative sizes always apply after we really have the proper width.

With this, we set the minimum size back to 180px, which is what it was before #9287.