Require 'package' instead of 'package/' so webpack activates sharing
This addresses the underlying problem in at least part of #9289.

Webpack module federation was not creating shared modules for the direct extensions, only for their dependencies. For example, @jupyterlab/application was packaged as a shared module, but @jupyterlab/application-extension was not.

The reasoning here is a bit subtle. The problem came down to how these top-level extensions were imported. The template made the require statements similar to @jupyterlab/application-extension/, with a trailing slash (which was nominally a slash between the package name and the actual import in the package, but mostly that actual import was an empty string, so we’d just be left with a trailing slash). However, webpack interpreted that trailing slash as indicating the import was only for paths starting with the package name, and not as a top-level package import. This meant that webpack did not see that these modules were being used, so did not include them in the list of modules packaged as shared modules. In other words, these top-level imports were treated as this case:

https://github.com/webpack/webpack/blob/7415a618460795874fdf95c8f03363ea96709d52/lib/sharing/ProvideSharedPlugin.js#L97-L99

instead of as this case:

https://github.com/webpack/webpack/blob/7415a618460795874fdf95c8f03363ea96709d52/lib/sharing/ProvideSharedPlugin.js#L100-L102

The fix here is simple - only put the slash in the require statement if we actually have a path to import.