Make built-in extensions shared eagerly
This ensures that the extensions can be packed into larger bundles, rather than individually bundled. The assumption here is that built-in things will all be needed, or at least that relatively few will be disabled.

Individual extensions can still load parts of themselves asynchronously to be separately bundled and loaded on-demand (for example, codemirror themes, the codemirror vim extension, etc.).