Update dependency graph script.
I took this out of the documentation since:

1. It has to be run manually, so often is out of date.
2. It gives the transitive reduction of the dependency graph only between non-extension jlab packages, which is not that helpful to understand relationships between plugins.

It still can be useful to understand relationships between library packages, so I didn’t delete the script entirely. If we can find a way to automatically generate it to keep it up to date, I think it would still be helpful (with appropriate caveats) in the JupyterLab development documentation.